string: add strtok/strtokv
Add an implementation of libc's standard strtok(3), which is useful
for tokenizing strings.

Also, add a version that will collect all tokens from a string into an
array, which is useful in situations where you need to know how many
tokens there are, and when a token's relative position in the order is
significant.